requirements.txt
requests
anthropic
pdfplumber
pymupdf
fastapi
uvicorn
python-multipart
python-dotenv
streamlit
pandas